feat(swiftui home): some refactoring/code optimization
